requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==3.2.1
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
logzero==1.7.0
packaging==24.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
smartapi-python==1.4.8
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
whitenoise==6.6.0